Code inspection

+++ app/src/main/java/ch/bailu/aat/services/background/DownloadStatistics.java
@@ -58,7 +58,7 @@ public class DownloadStatistics {
 
         if (success > 0) {
             builder.append("<br>Average file pixelCount: ");
-            builder.append(Math.round(bytes / success));
+            builder.append(Math.round(bytes / (float)success));
             builder.append(" bytes");
         }
